requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.6
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
hyperframe==6.0.1
idna==3.10
jiter==0.8.2
loguru==0.7.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
nonebot-adapter-mirai==2.3.3
nonebot-adapter-onebot==2.4.6
nonebot-adapter-qq==1.5.3
openai==1.57.1
propcache==0.2.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pygtrie==2.5.0
pyOpenSSL==24.3.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
setuptools==75.6.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
watchfiles==1.0.0
websockets==14.1
win32_setctime==1.2.0
yarl==1.18.3
